[refactor] migrate icarl update_memory to base learner
